Черновой редизайн профиля
- Исправлено отображение ошибок в продакшене
- Расширена тестовая страница до страницы профиля
- Выполнен небольшой редизайн и рефакторинг
