+++ frontend/Money.Web/Layout/NavMenu.razor
@@ -49,7 +49,7 @@
                 Настройки
             </MudNavLink>
 
-            <MudNavLink Href="auth"
+            <MudNavLink Href="profile"
                         Icon="@Icons.Material.Rounded.Person"
                         Match="NavLinkMatch.Prefix">
                 @context.User.Identity?.Name
