--- frontend/Money.Web/Pages/Account/Auth.razor
@@ -1,60 +1,0 @@
-﻿@page "/auth"
-@attribute [Authorize]
-
-<PageTitle>Добро пожаловать!</PageTitle>
-
-<MudText GutterBottom Typo="Typo.h3">
-    Вы успешно вошли в систему
-</MudText>
-
-<AuthorizeView>
-    <MudText Class="mb-4">Здравствуйте, @context.User.Identity?.Name!</MudText>
-</AuthorizeView>
-
-@if (_emailConfirmed)
-{
-    <MudText>Email-адрес подтвержден.</MudText>
-}
-else
-{
-    <MudPaper Class="pa-4" Elevation="2">
-        <MudText Typo="Typo.h6" GutterBottom>Подтвердите вашу почту</MudText>
-        <MudText Class="mb-4" Color="Color.Secondary">На почту @_email отправлен код подтверждения</MudText>
-
-        <EditForm Model="@_confirmationModel" OnValidSubmit="@HandleConfirmation">
-            <MudTextField @bind-Value="_confirmationModel.Code"
-                          Label="Код подтверждения"
-                          Required
-                          RequiredError="Введите код"
-                          Class="mb-4" />
-
-            <MudButton ButtonType="ButtonType.Submit"
-                       Color="Color.Primary"
-                       Class="mr-2"
-                       Variant="Variant.Filled"
-                       Disabled="@_isProcessing">
-                Подтвердить
-            </MudButton>
-
-            <MudButton Color="Color.Secondary"
-                       Variant="Variant.Text"
-                       Disabled="@(!_canResend)"
-                       OnClick="@ResendCode">
-                Отправить код повторно @(_remainingTime > 0 ? $"({_remainingTime}s)" : "")
-            </MudButton>
-        </EditForm>
-
-        @if (!string.IsNullOrEmpty(_errorMessage))
-        {
-            <MudAlert Class="mt-4" Icon="@Icons.Material.Outlined.Error" Severity="Severity.Error">
-                @_errorMessage
-            </MudAlert>
-        }
-    </MudPaper>
-}
-
-<MudButton Color="Color.Primary"
-           Href="/"
-           Variant="Variant.Filled">
-    Перейти на главную
-</MudButton>
